Make serde optional / move types out of near-network-primites back to near-network
