+++ chain/client/src/sync.rs
@@ -1218,13 +1218,14 @@ mod test {
     use near_chain::{ChainGenesis, Provenance};
     use near_crypto::{KeyType, PublicKey};
     use near_network::test_utils::MockNetworkAdapter;
-    use near_network::types::{EdgeInfo, PeerChainInfoV2};
+    use near_network::types::PeerChainInfoV2;
     use near_network::PeerInfo;
     use near_primitives::block::{Approval, Block, GenesisId};
     use near_primitives::network::PeerId;
 
     use super::*;
     use crate::test_utils::TestEnv;
+    use near_network::routing::EdgeInfo;
     use near_primitives::merkle::PartialMerkleTree;
     use near_primitives::types::EpochId;
     use near_primitives::validator_signer::InMemoryValidatorSigner;
