+++ chain/jsonrpc-primitives/Cargo.toml
@@ -22,7 +22,7 @@ near-metrics = { path = "../../core/metrics" }
 near-primitives = { path = "../../core/primitives" }
 near-primitives-core = { path = "../../core/primitives-core" }
 near-rpc-error-macro = { path = "../../tools/rpctypegen/macro" }
-near-network-primitives = { path = "../network-primitives", optional=true }
+near-network = { path = "../network", optional=true }
 
 [features]
-adversarial = ["near-network-primitives"]
+adversarial = ["near-network"]
