+++ chain/network-primitives/src/types.rs
@@ -2,7 +2,7 @@ use std::collections::{HashMap, HashSet};
 use std::convert::{Into, TryFrom};
 use std::fmt;
 use std::fmt::{Debug, Error, Formatter};
-use std::hash::{Hash, Hasher};
+use std::hash::Hash;
 use std::net::{AddrParseError, IpAddr, SocketAddr};
 use std::str::FromStr;
 use std::time::Duration;
@@ -10,7 +10,6 @@ use std::time::Duration;
 use actix::dev::{MessageResponse, ResponseChannel};
 use actix::{Actor, Message};
 use borsh::{BorshDeserialize, BorshSerialize};
-use byteorder::{LittleEndian, WriteBytesExt};
 use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
 use strum::AsStaticStr;
@@ -33,9 +32,7 @@ use near_primitives::syncing::{
     ShardStateSyncResponseV1,
 };
 use near_primitives::transaction::{ExecutionOutcomeWithIdAndProof, SignedTransaction};
-use near_primitives::types::{
-    AccountId, BlockHeight, BlockReference, EpochId, ProtocolVersion, ShardId,
-};
+use near_primitives::types::{AccountId, BlockHeight, BlockReference, EpochId, ShardId};
 use near_primitives::utils::{from_timestamp, to_timestamp};
 use near_primitives::views::{FinalExecutionOutcomeView, QueryRequest, QueryResponse};
 
@@ -1022,345 +1019,6 @@ impl PartialEncodedChunkForwardMsg {
         let correct_hash = combine_hash(self.inner_header_hash, self.merkle_root);
         ChunkHash(correct_hash) == self.chunk_hash
     }
-}
-
-/// Represents edge between two nodes. Unlike `Edge` it doesn't contain signatures.
-#[derive(Hash, Clone, Eq, PartialEq, Debug, Serialize, Deserialize)]
-pub struct SimpleEdge {
-    key: (PeerId, PeerId),
-    nonce: u64,
-}
-
-impl SimpleEdge {
-    pub fn new(peer0: PeerId, peer1: PeerId, nonce: u64) -> SimpleEdge {
-        let (peer0, peer1) = Edge::key(peer0, peer1);
-        SimpleEdge { key: (peer0, peer1), nonce }
-    }
-
-    pub fn key(&self) -> &(PeerId, PeerId) {
-        &self.key
-    }
-
-    pub fn nonce(&self) -> u64 {
-        self.nonce
-    }
-
-    pub fn edge_type(&self) -> EdgeType {
-        if self.nonce % 2 == 1 {
-            EdgeType::Added
-        } else {
-            EdgeType::Removed
-        }
-    }
-}
-
-/// Status of the edge
-#[derive(BorshSerialize, BorshDeserialize, Clone, PartialEq, Eq, Debug, Hash)]
-pub enum EdgeType {
-    Added,
-    Removed,
-}
-
-/// Edge object. Contains information relative to a new edge that is being added or removed
-/// from the network. This is the information that is required.
-#[derive(BorshSerialize, BorshDeserialize, Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-pub struct Edge {
-    /// Since edges are not directed `peer0 < peer1` should hold.
-    pub peer0: PeerId,
-    pub peer1: PeerId,
-    /// Nonce to keep tracking of the last update on this edge.
-    /// It must be even
-    pub nonce: u64,
-    /// Signature from parties validating the edge. These are signature of the added edge.
-    signature0: Signature,
-    signature1: Signature,
-    /// Info necessary to declare an edge as removed.
-    /// The bool says which party is removing the edge: false for Peer0, true for Peer1
-    /// The signature from the party removing the edge.
-    removal_info: Option<(bool, Signature)>,
-}
-
-impl Hash for Edge {
-    fn hash<H: Hasher>(&self, state: &mut H) {
-        state.write(&self.peer0.0.try_to_vec().unwrap());
-        state.write(&self.peer1.0.try_to_vec().unwrap());
-        state.write_u64(self.nonce)
-    }
-}
-
-impl Edge {
-    /// Create an addition edge.
-    pub fn new(
-        peer0: PeerId,
-        peer1: PeerId,
-        nonce: u64,
-        signature0: Signature,
-        signature1: Signature,
-    ) -> Self {
-        let (peer0, signature0, peer1, signature1) = if peer0 < peer1 {
-            (peer0, signature0, peer1, signature1)
-        } else {
-            (peer1, signature1, peer0, signature0)
-        };
-
-        Self { peer0, peer1, nonce, signature0, signature1, removal_info: None }
-    }
-
-    pub fn to_simple_edge(&self) -> SimpleEdge {
-        SimpleEdge::new(self.peer0.clone(), self.peer1.clone(), self.nonce)
-    }
-
-    pub fn make_fake_edge(peer0: PeerId, peer1: PeerId, nonce: u64) -> Self {
-        Self {
-            peer0,
-            peer1,
-            nonce,
-            signature0: Signature::empty(KeyType::ED25519),
-            signature1: Signature::empty(KeyType::ED25519),
-            removal_info: None,
-        }
-    }
-
-    /// Build a new edge with given information from the other party.
-    pub fn build_with_secret_key(
-        peer0: PeerId,
-        peer1: PeerId,
-        nonce: u64,
-        secret_key: &SecretKey,
-        signature1: Signature,
-    ) -> Self {
-        let hash = if peer0 < peer1 {
-            Edge::build_hash(&peer0, &peer1, nonce)
-        } else {
-            Edge::build_hash(&peer1, &peer0, nonce)
-        };
-        let signature0 = secret_key.sign(hash.as_ref());
-        Edge::new(peer0, peer1, nonce, signature0, signature1)
-    }
-
-    /// Create the remove edge change from an added edge change.
-    pub fn remove_edge(&self, me: PeerId, sk: &SecretKey) -> Self {
-        assert_eq!(self.edge_type(), EdgeType::Added);
-        let mut edge = self.clone();
-        edge.nonce += 1;
-        let me = edge.peer0 == me;
-        let hash = edge.hash();
-        let signature = sk.sign(hash.as_ref());
-        edge.removal_info = Some((me, signature));
-        edge
-    }
-
-    /// Build the hash of the edge given its content.
-    /// It is important that peer0 < peer1 at this point.
-    fn build_hash(peer0: &PeerId, peer1: &PeerId, nonce: u64) -> CryptoHash {
-        let mut buffer = Vec::<u8>::new();
-        let peer0: Vec<u8> = peer0.clone().into();
-        buffer.extend_from_slice(peer0.as_slice());
-        let peer1: Vec<u8> = peer1.clone().into();
-        buffer.extend_from_slice(peer1.as_slice());
-        buffer.write_u64::<LittleEndian>(nonce).unwrap();
-        hash(buffer.as_slice())
-    }
-
-    fn hash(&self) -> CryptoHash {
-        Edge::build_hash(&self.peer0, &self.peer1, self.nonce)
-    }
-
-    fn prev_hash(&self) -> CryptoHash {
-        Edge::build_hash(&self.peer0, &self.peer1, self.nonce - 1)
-    }
-
-    pub fn verify(&self) -> bool {
-        if self.peer0 > self.peer1 {
-            return false;
-        }
-
-        match self.edge_type() {
-            EdgeType::Added => {
-                let data = self.hash();
-
-                self.removal_info.is_none()
-                    && self.signature0.verify(data.as_ref(), &self.peer0.public_key())
-                    && self.signature1.verify(data.as_ref(), &self.peer1.public_key())
-            }
-            EdgeType::Removed => {
-                // nonce should be an even positive number
-                if self.nonce == 0 {
-                    return false;
-                }
-
-                // Check referring added edge is valid.
-                let add_hash = self.prev_hash();
-                if !self.signature0.verify(add_hash.as_ref(), &self.peer0.public_key())
-                    || !self.signature1.verify(add_hash.as_ref(), &self.peer1.public_key())
-                {
-                    return false;
-                }
-
-                if let Some((party, signature)) = &self.removal_info {
-                    let peer = if *party { &self.peer0 } else { &self.peer1 };
-                    let del_hash = self.hash();
-                    signature.verify(del_hash.as_ref(), &peer.public_key())
-                } else {
-                    false
-                }
-            }
-        }
-    }
-
-    pub fn key(peer0: PeerId, peer1: PeerId) -> (PeerId, PeerId) {
-        if peer0 < peer1 {
-            (peer0, peer1)
-        } else {
-            (peer1, peer0)
-        }
-    }
-
-    /// Helper function when adding a new edge and we receive information from new potential peer
-    /// to verify the signature.
-    pub fn partial_verify(peer0: PeerId, peer1: PeerId, edge_info: &EdgeInfo) -> bool {
-        let pk = peer1.public_key();
-        let (peer0, peer1) = Edge::key(peer0, peer1);
-        let data = Edge::build_hash(&peer0, &peer1, edge_info.nonce);
-        edge_info.signature.verify(data.as_ref(), &pk)
-    }
-
-    pub fn get_pair(&self) -> (PeerId, PeerId) {
-        (self.peer0.clone(), self.peer1.clone())
-    }
-
-    /// It will be considered as a new edge if the nonce is odd, otherwise it is canceling the
-    /// previous edge.
-    pub fn edge_type(&self) -> EdgeType {
-        if self.nonce % 2 == 1 {
-            EdgeType::Added
-        } else {
-            EdgeType::Removed
-        }
-    }
-
-    /// Next nonce of valid addition edge.
-    pub fn next_nonce(nonce: u64) -> u64 {
-        if nonce % 2 == 1 {
-            nonce + 2
-        } else {
-            nonce + 1
-        }
-    }
-
-    /// Next nonce of valid addition edge.
-    pub fn next(&self) -> u64 {
-        Edge::next_nonce(self.nonce)
-    }
-
-    pub fn contains_peer(&self, peer_id: &PeerId) -> bool {
-        self.peer0 == *peer_id || self.peer1 == *peer_id
-    }
-
-    /// Find a peer id in this edge different from `me`.
-    pub fn other(&self, me: &PeerId) -> Option<PeerId> {
-        if self.peer0 == *me {
-            Some(self.peer1.clone())
-        } else if self.peer1 == *me {
-            Some(self.peer0.clone())
-        } else {
-            None
-        }
-    }
-}
-
-/// Information that will be ultimately used to create a new edge.
-/// It contains nonce proposed for the edge with signature from peer.
-#[derive(Clone, BorshSerialize, BorshDeserialize, PartialEq, Eq, Debug, Default)]
-pub struct EdgeInfo {
-    pub nonce: u64,
-    pub signature: Signature,
-}
-
-impl EdgeInfo {
-    pub fn new(peer0: PeerId, peer1: PeerId, nonce: u64, secret_key: &SecretKey) -> Self {
-        let (peer0, peer1) = Edge::key(peer0, peer1);
-        let data = Edge::build_hash(&peer0, &peer1, nonce);
-        let signature = secret_key.sign(data.as_ref());
-        Self { nonce, signature }
-    }
-}
-
-#[derive(MessageResponse, Debug, Serialize)]
-pub struct GetPeerIdResult {
-    pub peer_id: PeerId,
-}
-
-#[derive(Debug, Serialize)]
-pub struct ActivePeerInfo {
-    pub peer_info: PeerInfo,
-    pub protocol_version: ProtocolVersion,
-    /// Number of bytes we've received from the peer.
-    pub received_bytes_per_sec: u64,
-    /// Number of bytes we've sent to the peer.
-    pub sent_bytes_per_sec: u64,
-    /// Who started connection. Inbound (other) or Outbound (us).
-    pub peer_type: PeerType,
-}
-
-#[derive(
-    BorshSerialize,
-    BorshDeserialize,
-    Serialize,
-    Deserialize,
-    PartialOrd,
-    Ord,
-    PartialEq,
-    Eq,
-    Copy,
-    Clone,
-    Debug,
-)]
-
-pub struct ValidIBFLevel(pub u64);
-
-/// We create IbfSets of various sizes from 2^10+2 up to 2^17+2. Those constants specify valid ranges.
-pub const MIN_IBF_LEVEL: ValidIBFLevel = ValidIBFLevel(10);
-pub const MAX_IBF_LEVEL: ValidIBFLevel = ValidIBFLevel(17);
-
-/// Represents IbfLevel from 10 to 17.
-impl ValidIBFLevel {
-    pub fn inc(&self) -> Option<ValidIBFLevel> {
-        if self.0 + 1 >= MIN_IBF_LEVEL.0 && self.0 + 1 <= MAX_IBF_LEVEL.0 {
-            Some(ValidIBFLevel(self.0 + 1))
-        } else {
-            None
-        }
-    }
-
-    pub fn is_valid(&self) -> bool {
-        return self.0 >= MIN_IBF_LEVEL.0 && self.0 <= MAX_IBF_LEVEL.0;
-    }
-}
-
-#[derive(Serialize)]
-pub struct PeerRequestResult {
-    pub peers: Vec<PeerInfo>,
-}
-
-impl<A, M> MessageResponse<A, M> for PeerRequestResult
-where
-    A: Actor,
-    M: Message<Result = PeerRequestResult>,
-{
-    fn handle<R: ResponseChannel<M>>(self, _: &mut A::Context, tx: Option<R>) {
-        if let Some(tx) = tx {
-            tx.send(self)
-        }
-    }
-}
-
-#[derive(MessageResponse, Debug, Serialize)]
-pub struct SetAdvOptionsResult {}
-
-#[derive(MessageResponse, Debug, Serialize)]
-pub struct GetRoutingTableResult {
-    pub edges_info: Vec<SimpleEdge>,
 }
 
 #[cfg(test)]
