+++ chain/network/Cargo.toml
@@ -20,6 +20,7 @@ strum = { version = "0.20", features = ["derive"] }
 near-rust-allocator-proxy = "0.2.9"
 bytesize = "1.0.1"
 conqueue = "0.4.0"
+serde = { version = "1", features = ["derive"], optional=true }
 
 borsh = "0.9"
 cached = "0.23"
@@ -44,7 +45,7 @@ tempfile = "3"
 bencher = "0.1.5"
 
 [features]
-adversarial = ["near-network-primitives/adversarial"]
+adversarial = ["near-network-primitives/adversarial", "serde"]
 delay_detector = ["delay-detector"]
 performance_stats = ["near-performance-metrics/performance_stats"]
 sandbox = ["near-network-primitives/sandbox"]
