+++ chain/network/src/routing.rs
@@ -6,13 +6,14 @@ use std::time::Instant;
 use cached::{Cached, SizedCache};
 use chrono;
 use conqueue::{QueueReceiver, QueueSender};
+#[cfg(feature = "adversarial")]
+use serde::{Deserialize, Serialize};
 use tracing::{debug, trace, warn};
 
 #[cfg(feature = "delay_detector")]
 use delay_detector::DelayDetector;
 use near_metrics;
-use near_network_primitives::types::SimpleEdge;
-use near_primitives::hash::CryptoHash;
+use near_primitives::hash::{hash, CryptoHash};
 use near_primitives::network::{AnnounceAccount, PeerId};
 use near_primitives::types::AccountId;
 use near_primitives::utils::index_to_bytes;
@@ -21,13 +22,18 @@ use near_store::{
     StoreUpdate,
 };
 
-use crate::metrics;
-use crate::types::{Edge, EdgeType};
 use crate::{
     cache::RouteBackCache,
     types::{PeerIdOrHash, Ping, Pong},
     utils::cache_to_hashmap,
 };
+use crate::{metrics, PeerInfo};
+use actix::dev::{MessageResponse, ResponseChannel};
+use actix::{Actor, Message};
+use borsh::{BorshDeserialize, BorshSerialize};
+use byteorder::{LittleEndian, WriteBytesExt};
+use near_crypto::{KeyType, SecretKey, Signature};
+use std::hash::{Hash, Hasher};
 
 const ANNOUNCE_ACCOUNT_CACHE_SIZE: usize = 10_000;
 const ROUTE_BACK_CACHE_SIZE: u64 = 100_000;
@@ -43,6 +49,319 @@ pub const SAVE_PEERS_MAX_TIME: u64 = 7_200;
 pub const SAVE_PEERS_AFTER_TIME: u64 = 3_600;
 /// Graph implementation supports up to 128 peers.
 pub const MAX_NUM_PEERS: usize = 128;
+
+/// Information that will be ultimately used to create a new edge.
+/// It contains nonce proposed for the edge with signature from peer.
+#[derive(Clone, BorshSerialize, BorshDeserialize, PartialEq, Eq, Debug, Default)]
+pub struct EdgeInfo {
+    pub nonce: u64,
+    pub signature: Signature,
+}
+
+impl EdgeInfo {
+    pub fn new(peer0: PeerId, peer1: PeerId, nonce: u64, secret_key: &SecretKey) -> Self {
+        let (peer0, peer1) = Edge::key(peer0, peer1);
+        let data = Edge::build_hash(&peer0, &peer1, nonce);
+        let signature = secret_key.sign(data.as_ref());
+        Self { nonce, signature }
+    }
+}
+
+/// Status of the edge
+#[derive(BorshSerialize, BorshDeserialize, Clone, PartialEq, Eq, Debug, Hash)]
+pub enum EdgeType {
+    Added,
+    Removed,
+}
+
+/// Edge object. Contains information relative to a new edge that is being added or removed
+/// from the network. This is the information that is required.
+#[derive(BorshSerialize, BorshDeserialize, Clone, Debug, PartialEq, Eq)]
+#[cfg_attr(feature = "adversarial", derive(Serialize, Deserialize))]
+pub struct Edge {
+    /// Since edges are not directed `peer0 < peer1` should hold.
+    pub peer0: PeerId,
+    pub peer1: PeerId,
+    /// Nonce to keep tracking of the last update on this edge.
+    /// It must be even
+    pub nonce: u64,
+    /// Signature from parties validating the edge. These are signature of the added edge.
+    signature0: Signature,
+    signature1: Signature,
+    /// Info necessary to declare an edge as removed.
+    /// The bool says which party is removing the edge: false for Peer0, true for Peer1
+    /// The signature from the party removing the edge.
+    removal_info: Option<(bool, Signature)>,
+}
+
+impl Hash for Edge {
+    fn hash<H: Hasher>(&self, state: &mut H) {
+        state.write(&self.peer0.0.try_to_vec().unwrap());
+        state.write(&self.peer1.0.try_to_vec().unwrap());
+        state.write_u64(self.nonce)
+    }
+}
+
+impl Edge {
+    /// Create an addition edge.
+    pub fn new(
+        peer0: PeerId,
+        peer1: PeerId,
+        nonce: u64,
+        signature0: Signature,
+        signature1: Signature,
+    ) -> Self {
+        let (peer0, signature0, peer1, signature1) = if peer0 < peer1 {
+            (peer0, signature0, peer1, signature1)
+        } else {
+            (peer1, signature1, peer0, signature0)
+        };
+
+        Self { peer0, peer1, nonce, signature0, signature1, removal_info: None }
+    }
+
+    pub fn to_simple_edge(&self) -> SimpleEdge {
+        SimpleEdge::new(self.peer0.clone(), self.peer1.clone(), self.nonce)
+    }
+
+    pub fn make_fake_edge(peer0: PeerId, peer1: PeerId, nonce: u64) -> Self {
+        Self {
+            peer0,
+            peer1,
+            nonce,
+            signature0: Signature::empty(KeyType::ED25519),
+            signature1: Signature::empty(KeyType::ED25519),
+            removal_info: None,
+        }
+    }
+
+    /// Build a new edge with given information from the other party.
+    pub fn build_with_secret_key(
+        peer0: PeerId,
+        peer1: PeerId,
+        nonce: u64,
+        secret_key: &SecretKey,
+        signature1: Signature,
+    ) -> Self {
+        let hash = if peer0 < peer1 {
+            Edge::build_hash(&peer0, &peer1, nonce)
+        } else {
+            Edge::build_hash(&peer1, &peer0, nonce)
+        };
+        let signature0 = secret_key.sign(hash.as_ref());
+        Edge::new(peer0, peer1, nonce, signature0, signature1)
+    }
+
+    /// Create the remove edge change from an added edge change.
+    pub fn remove_edge(&self, me: PeerId, sk: &SecretKey) -> Self {
+        assert_eq!(self.edge_type(), EdgeType::Added);
+        let mut edge = self.clone();
+        edge.nonce += 1;
+        let me = edge.peer0 == me;
+        let hash = edge.hash();
+        let signature = sk.sign(hash.as_ref());
+        edge.removal_info = Some((me, signature));
+        edge
+    }
+
+    /// Build the hash of the edge given its content.
+    /// It is important that peer0 < peer1 at this point.
+    fn build_hash(peer0: &PeerId, peer1: &PeerId, nonce: u64) -> CryptoHash {
+        let mut buffer = Vec::<u8>::new();
+        let peer0: Vec<u8> = peer0.clone().into();
+        buffer.extend_from_slice(peer0.as_slice());
+        let peer1: Vec<u8> = peer1.clone().into();
+        buffer.extend_from_slice(peer1.as_slice());
+        buffer.write_u64::<LittleEndian>(nonce).unwrap();
+        hash(buffer.as_slice())
+    }
+
+    fn hash(&self) -> CryptoHash {
+        Edge::build_hash(&self.peer0, &self.peer1, self.nonce)
+    }
+
+    fn prev_hash(&self) -> CryptoHash {
+        Edge::build_hash(&self.peer0, &self.peer1, self.nonce - 1)
+    }
+
+    pub fn verify(&self) -> bool {
+        if self.peer0 > self.peer1 {
+            return false;
+        }
+
+        match self.edge_type() {
+            EdgeType::Added => {
+                let data = self.hash();
+
+                self.removal_info.is_none()
+                    && self.signature0.verify(data.as_ref(), &self.peer0.public_key())
+                    && self.signature1.verify(data.as_ref(), &self.peer1.public_key())
+            }
+            EdgeType::Removed => {
+                // nonce should be an even positive number
+                if self.nonce == 0 {
+                    return false;
+                }
+
+                // Check referring added edge is valid.
+                let add_hash = self.prev_hash();
+                if !self.signature0.verify(add_hash.as_ref(), &self.peer0.public_key())
+                    || !self.signature1.verify(add_hash.as_ref(), &self.peer1.public_key())
+                {
+                    return false;
+                }
+
+                if let Some((party, signature)) = &self.removal_info {
+                    let peer = if *party { &self.peer0 } else { &self.peer1 };
+                    let del_hash = self.hash();
+                    signature.verify(del_hash.as_ref(), &peer.public_key())
+                } else {
+                    false
+                }
+            }
+        }
+    }
+
+    pub fn key(peer0: PeerId, peer1: PeerId) -> (PeerId, PeerId) {
+        if peer0 < peer1 {
+            (peer0, peer1)
+        } else {
+            (peer1, peer0)
+        }
+    }
+
+    /// Helper function when adding a new edge and we receive information from new potential peer
+    /// to verify the signature.
+    pub fn partial_verify(peer0: PeerId, peer1: PeerId, edge_info: &EdgeInfo) -> bool {
+        let pk = peer1.public_key();
+        let (peer0, peer1) = Edge::key(peer0, peer1);
+        let data = Edge::build_hash(&peer0, &peer1, edge_info.nonce);
+        edge_info.signature.verify(data.as_ref(), &pk)
+    }
+
+    pub fn get_pair(&self) -> (PeerId, PeerId) {
+        (self.peer0.clone(), self.peer1.clone())
+    }
+
+    /// It will be considered as a new edge if the nonce is odd, otherwise it is canceling the
+    /// previous edge.
+    pub fn edge_type(&self) -> EdgeType {
+        if self.nonce % 2 == 1 {
+            EdgeType::Added
+        } else {
+            EdgeType::Removed
+        }
+    }
+
+    /// Next nonce of valid addition edge.
+    pub fn next_nonce(nonce: u64) -> u64 {
+        if nonce % 2 == 1 {
+            nonce + 2
+        } else {
+            nonce + 1
+        }
+    }
+
+    /// Next nonce of valid addition edge.
+    pub fn next(&self) -> u64 {
+        Edge::next_nonce(self.nonce)
+    }
+
+    pub fn contains_peer(&self, peer_id: &PeerId) -> bool {
+        self.peer0 == *peer_id || self.peer1 == *peer_id
+    }
+
+    /// Find a peer id in this edge different from `me`.
+    pub fn other(&self, me: &PeerId) -> Option<PeerId> {
+        if self.peer0 == *me {
+            Some(self.peer1.clone())
+        } else if self.peer1 == *me {
+            Some(self.peer0.clone())
+        } else {
+            None
+        }
+    }
+}
+
+/// Represents edge between two nodes. Unlike `Edge` it doesn't contain signatures.
+#[derive(Hash, Clone, Eq, PartialEq, Debug)]
+#[cfg_attr(feature = "adversarial", derive(Serialize, Deserialize))]
+pub struct SimpleEdge {
+    key: (PeerId, PeerId),
+    nonce: u64,
+}
+
+impl SimpleEdge {
+    pub fn new(peer0: PeerId, peer1: PeerId, nonce: u64) -> SimpleEdge {
+        let (peer0, peer1) = Edge::key(peer0, peer1);
+        SimpleEdge { key: (peer0, peer1), nonce }
+    }
+
+    pub fn key(&self) -> &(PeerId, PeerId) {
+        &self.key
+    }
+
+    pub fn nonce(&self) -> u64 {
+        self.nonce
+    }
+
+    pub fn edge_type(&self) -> EdgeType {
+        if self.nonce % 2 == 1 {
+            EdgeType::Added
+        } else {
+            EdgeType::Removed
+        }
+    }
+}
+
+#[derive(BorshSerialize, BorshDeserialize, PartialEq, Eq, Clone, Debug, Copy)]
+pub struct ValidIBFLevel(pub u64);
+
+/// We create IbfSets of various sizes from 2^10+2 up to 2^17+2. Those constants specify valid ranges.
+pub const MIN_IBF_LEVEL: ValidIBFLevel = ValidIBFLevel(10);
+pub const MAX_IBF_LEVEL: ValidIBFLevel = ValidIBFLevel(17);
+
+/// Represents IbfLevel from 10 to 17.
+impl ValidIBFLevel {
+    pub fn inc(&self) -> Option<ValidIBFLevel> {
+        if self.0 + 1 >= MIN_IBF_LEVEL.0 && self.0 + 1 <= MAX_IBF_LEVEL.0 {
+            Some(ValidIBFLevel(self.0 + 1))
+        } else {
+            None
+        }
+    }
+
+    pub fn is_valid(&self) -> bool {
+        return self.0 >= MIN_IBF_LEVEL.0 && self.0 <= MAX_IBF_LEVEL.0;
+    }
+}
+
+#[cfg_attr(feature = "adversarial", derive(Serialize))]
+pub struct PeerRequestResult {
+    pub peers: Vec<PeerInfo>,
+}
+
+impl<A, M> MessageResponse<A, M> for PeerRequestResult
+where
+    A: Actor,
+    M: Message<Result = PeerRequestResult>,
+{
+    fn handle<R: ResponseChannel<M>>(self, _: &mut A::Context, tx: Option<R>) {
+        if let Some(tx) = tx {
+            tx.send(self)
+        }
+    }
+}
+
+#[derive(MessageResponse, Debug)]
+#[cfg_attr(feature = "adversarial", derive(Serialize))]
+pub struct SetAdvOptionsResult {}
+
+#[derive(MessageResponse, Debug)]
+#[cfg_attr(feature = "adversarial", derive(Serialize))]
+pub struct GetRoutingTableResult {
+    pub edges_info: Vec<SimpleEdge>,
+}
 
 pub struct EdgeVerifierHelper {
     /// Shared version of edges_info used by multiple threads
