+++ chain/network/tests/cache_edges.rs
@@ -5,9 +5,10 @@ use borsh::de::BorshDeserialize;
 use chrono::{DateTime, Duration, Utc};
 
 use near_crypto::Signature;
-use near_network::routing::{RoutingTable, SAVE_PEERS_AFTER_TIME, SAVE_PEERS_MAX_TIME};
+use near_network::routing::{
+    Edge, EdgeType, RoutingTable, SAVE_PEERS_AFTER_TIME, SAVE_PEERS_MAX_TIME,
+};
 use near_network::test_utils::random_peer_id;
-use near_network_primitives::types::{Edge, EdgeType};
 use near_primitives::network::PeerId;
 use near_store::test_utils::create_test_store;
 use near_store::{ColComponentEdges, ColPeerComponent, Store};
